backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
anthropic==0.28.0
deepgram-sdk==3.4.0
faiss-cpu==1.8.0
networkx==3.3
librosa==0.10.2
numpy==1.26.4
websockets==12.0
python-multipart==0.0.9
httpx==0.27.0